New features
- path and name of metadata file of unstitched datasets (mdata.bin) can be changed using command line option --mdata_bin
- loading of special cases of unstitched subvolumes has been optimized

+++ src/imagemanager/VirtualVolume.h
@@ -88,7 +88,7 @@ protected:
 				src += totalChanSize;
 				cc++;
 			}
-			memcpy(dst,src,totalChanSize);
+			memcpy(dst,src,(size_t)totalChanSize);
 		}
 	}
